Correct type-casting bug in vgrid::C_genab_5002_5003()
- The bug expressed itself when changing compilers from intel16 to
  intel19, and when changing hosts from ppp1/2 to ppp3/4
- This allows the test, c_new_gen_all, to pass (partially) for the
  a_m_8 array for the dm_5002_from_model_run file

+++ src/vgrid.cpp
@@ -19,7 +19,6 @@
  */
 
 #include "vgrid.hpp"
-#include "vgrid_descriptor.h"
 #include <stdlib.h>
 #include <stdio.h>
 #include <string.h>
@@ -5038,7 +5037,7 @@ int vgrid::C_genab_5002_5003(float *hybuser, int nk, int *nl_m, int *nl_t, float
   // Momentum levels
   pr1 = 1. / (zsrf_8 - ztop_8);
   for( k = 0; k < nk; k++ ) {
-    zeta_8  = zsrf_8 + log(hybuser[k]);
+    zeta_8  = zsrf_8 + log((double)hybuser[k]);
     lamba_8  = ( zeta_8 - ztop_8 ) * pr1;
     rcoef  = (float) (rcoef2 - ( rcoef2 - rcoef1 ) * lamba_8);
     b_m_8[k] = pow(lamba_8, rcoef);
